+++ Samples/Cpp/BuildExtension/Executable/Recipe.sml
@@ -1,7 +1,7 @@
 Name: "Samples.SimpleBuildExtension.Executable"
 Language: "C++|0.1"
 Type: "Executable"
-Version: "1.0.1"
+Version: "1.0.0"
 Source: [
 	"Main.cpp"
 ]
