+++ Source/GenerateSharp/Utilities/SML/SMLValue.cs
@@ -1,11 +1,8 @@
-﻿// <copyright file="SML.cs" company="Soup">
+﻿// <copyright file="SMLValue.cs" company="Soup">
 // Copyright (c) Soup. All rights reserved.
 // </copyright>
 
-using Soup.Build.Runtime;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace Soup.Build.Utilities
 {
@@ -18,150 +15,6 @@ namespace Soup.Build.Utilities
 		String,
 		Table,
 		Array,
-	}
-
-	public class SMLTable : IEquatable<SMLTable>
-	{
-		private Dictionary<string, SMLValue> _value;
-
-		public SMLTable()
-		{
-			_value = new Dictionary<string, SMLValue>();
-		}
-
-		public SMLTable(Dictionary<string, SMLValue> value)
-		{
-			_value = value;
-		}
-
-		public SMLValue this[string key]
-		{
-			get
-			{
-				return _value[key];
-			}
-		}
-
-		public Dictionary<string, SMLValue> GetValue()
-		{
-			return _value;
-		}
-
-		public override bool Equals(object? obj) => this.Equals(obj as SMLTable);
-
-		public bool Equals(SMLTable? rhs)
-		{
-			if (rhs is null)
-				return false;
-
-			// Optimization for a common success case.
-			if (object.ReferenceEquals(this, rhs))
-				return true;
-
-			// Return true if the fields match.
-			return Enumerable.SequenceEqual(this._value, rhs._value);
-		}
-
-		public override int GetHashCode() => (_value).GetHashCode();
-
-		public static bool operator ==(SMLTable? lhs, SMLTable? rhs)
-		{
-			if (lhs is null)
-			{
-				if (rhs is null)
-					return true;
-				else
-					return false;
-			}
-
-			return lhs.Equals(rhs);
-		}
-
-		public static bool operator !=(SMLTable? lhs, SMLTable? rhs) => !(lhs == rhs);
-	}
-
-	public class SMLArray : IEquatable<SMLArray>
-	{
-		private List<SMLValue> _value;
-
-		public SMLArray()
-		{
-			_value = new List<SMLValue>();
-		}
-
-		public SMLArray(List<SMLValue> value)
-
-		{
-			_value = value;
-		}
-
-		public SMLValue this[int key]
-		{
-			get
-			{
-				return _value[key];
-			}
-		}
-
-		public int GetSize()
-		{
-			return _value.Count;
-		}
-
-		public List<SMLValue> GetValue()
-		{
-			return _value;
-		}
-
-		public override bool Equals(object? obj) => this.Equals(obj as SMLArray);
-
-		public bool Equals(SMLArray? rhs)
-		{
-			if (rhs is null)
-				return false;
-
-			// Optimization for a common success case.
-			if (object.ReferenceEquals(this, rhs))
-				return true;
-
-			// Return true if the fields match.
-			return Enumerable.SequenceEqual(this._value, rhs._value);
-		}
-
-		public override int GetHashCode() => (_value).GetHashCode();
-
-		public static bool operator ==(SMLArray? lhs, SMLArray? rhs)
-		{
-			if (lhs is null)
-			{
-				if (rhs is null)
-					return true;
-				else
-					return false;
-			}
-
-			return lhs.Equals(rhs);
-		}
-
-		public static bool operator !=(SMLArray? lhs, SMLArray? rhs) => !(lhs == rhs);
-	}
-
-	/// <summary>
-	/// The SML Document
-	/// </summary>
-	public class SMLDocument
-	{
-		private SMLTable _root;
-
-		public SMLDocument(SMLTable root)
-		{
-			_root = root;
-		}
-
-		public SMLTable GetRoot()
-		{
-			return _root;
-		}
 	}
 
 	public class SMLValue : IEquatable<SMLValue>
